Update the PostgreSQL JDBC driver.
According to the official PostgreSQL JDBC driver web site[1] as of this date, the 9.1-901 version of the driver should be used for PostgreSQL version 7.2 and newer.  There is one driver for JVM 1.4 and 1.5 (JDBC3) and another for JVM 1.5 and 1.6 (JDBC4).  Both are now included, with the JDBC4 driver on the classpath.  To run with older JVMs, you may need to change the classpath.

There are dozens of contributors listed for the jars[2].  The jars are under BSD license with copyright held by the PostgreSQL Global Development Group[3].

[1] http://jdbc.postgresql.org/download.html#current
[2] http://jdbc.postgresql.org/who.html
[3] http://jdbc.postgresql.org/license.html

git-svn-id: http://oltpbenchmark.googlecode.com/svn/trunk@485 8c9255d6-0b70-036a-fcb6-8a999947b34c